python/requirements.txt
# Python dependencies for the quickstart example
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
